refactor: RestoreKey
